+++ packages/plugin-hardhat/src/type-extensions.ts
@@ -1,5 +1,6 @@
 import 'hardhat/types/runtime';
 
+import type { silenceWarnings } from '@openzeppelin/upgrades-core';
 import type { DeployFunction } from './deploy-proxy';
 import type { UpgradeFunction, PrepareUpgradeFunction } from './upgrade-proxy';
 
@@ -9,6 +10,7 @@ declare module 'hardhat/types/runtime' {
       deployProxy: DeployFunction;
       upgradeProxy: UpgradeFunction;
       prepareUpgrade: PrepareUpgradeFunction;
+      silenceWarnings: typeof silenceWarnings;
     };
   }
 }