+++ packages/plugin-truffle/src/index.ts
@@ -1,3 +1,5 @@
 export * from './deploy-proxy';
 export * from './upgrade-proxy';
 export { admin } from './admin';
+
+export { silenceWarnings } from '@openzeppelin/upgrades-core';